Japanese_MUD/requirements.txt
# Core requirements
python-dotenv==1.0.0
pillow==10.0.0

# OpenAI API for text generation
openai==1.3.0

# Requests for API calls
requests==2.31.0

# Terminal UI
prompt_toolkit>=3.0.0
colorama>=0.4.0
rich>=13.0.0
pyfiglet>=0.8.0

# Web Interface
flask>=2.0.0
flask-socketio>=5.0.0
eventlet>=0.30.0
simple-websocket>=0.10.0  # Additional websocket dependency

# Database
pymongo>=4.3.3
motor>=3.1.2

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0 